requirements.txt
python-telegram-bot
mysql-connector-python
nltk
pytest
Pillow